+++ controllers/moduleGroupsController.js
@@ -1,43 +1,67 @@
 const responseHelper = require('../helpers/responseHelper');
-const db = require('../database/database');
+const errorResponseHelper = require('../helpers/errorResponseHelper');
 const moduleGroupService = require('../services/moduleGroupService');
+const db = require('../database/database');
 
-exports.postModuleGroups = async (req, res) => {
-  let transaction = await db.sequelize.transaction();
+exports.postModuleGroups = async (req, res, next) => {
+  const transaction = await db.sequelize.transaction();
+
   try {
-    let moduleGroupToCreate = req.body;
-    let createdModuleGroup = await moduleGroupService.createModuleGroup(transaction, moduleGroupToCreate);
+    const ModuleGroupToCreate = req.body;
+    const createdModuleGroup = await moduleGroupService.createModuleGroup(transaction, ModuleGroupToCreate);
+
     transaction.commit();
     return responseHelper(res, 201, 'Successfully created', createdModuleGroup);
   } catch (error) {
     transaction.rollback();
-    return responseHelper(res, 500, 'Internal Server Error.');
+    return errorResponseHelper(res, next, error);
   }
 };
 
-exports.putModuleGroups = async (req, res) => {
-  let transaction = await db.sequelize.transaction();
-  const moduleGroupId = req.query.moduleGroupId;
+exports.putModuleGroups = async (req, res, next) => {
+  const moduleGroup_id = req.query.moduleGroupId;
+  const transaction = await db.sequelize.transaction();
+
   try {
-    let moduleGroupToUpdate = { ...req.body, moduleGroup_id: moduleGroupId };
-    let updatedModuleGroup = await moduleGroupService.updateModuleGroup(transaction, moduleGroupToUpdate);
+    if (!moduleGroup_id) {
+      throw new Error('No module group given');
+    }
+
+    const ModuleGroupToUpdate = req.body;
+    const updatedModuleGroup = await moduleGroupService.updateModuleGroup(transaction, {
+      ...ModuleGroupToUpdate,
+      moduleGroup_id,
+    });
+    if (!updatedModuleGroup) {
+      throw new Error('No module group to update');
+    }
+
     transaction.commit();
-    return responseHelper(res, 200, 'Successfully updated.', updatedModuleGroup);
+    return responseHelper(res, 200, 'Successfully updated', updatedModuleGroup);
   } catch (error) {
     transaction.rollback();
-    return responseHelper(res, 500, 'Internal Server Error.');
+    return errorResponseHelper(res, next, error);
   }
 };
 
-exports.deleteModuleGroups = async (req, res) => {
-  let transaction = await db.sequelize.transaction();
+exports.deleteModuleGroups = async (req, res, next) => {
   const moduleGroupId = req.query.moduleGroupId;
+  const transaction = await db.sequelize.transaction();
+
   try {
-    let deletedModuleGroup = await moduleGroupService.deleteModuleGroup(transaction, moduleGroupId);
+    if (!moduleGroupId) {
+      throw new Error('No module group given');
+    }
+
+    const deletedModulecatalog = await moduleGroupService.deleteModuleGroup(transaction, moduleGroupId);
+    if (!deletedModulecatalog) {
+      throw new Error('No module group found to delete');
+    }
+
     transaction.commit();
-    return responseHelper(res, 200, 'Successfully deleted', deletedModuleGroup);
+    return responseHelper(res, 200, 'Successfully deleted', deletedModulecatalog);
   } catch (error) {
     transaction.rollback();
-    return responseHelper(res, 500, 'Internal Server Error.');
+    return errorResponseHelper(res, next, error);
   }
 };
